backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.5.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
alembic==1.13.1
websockets==12.0
pytest==8.3.3
pytest-asyncio==0.24.0
